Load Creation Changes
Creating a new outgoing load now creates a new matching incoming load along with it. When going to create a new incoming load, it shows you the basic load information of incomplete outgoing loads for the user to choose which load is incoming. The user then enters all of the necessary information and that information gets filled into the already existing incoming load while then setting both to be complete.